requirements.txt
torch
faiss-cpu
numpy
transformers
sentence_transformers
pydantic
SentencePiece
fastapi
uvicorn
